Fix compilation error on Linux with HOST environment variable set
The Makefile variable "HOST" was introduced to cross-compile Tux Paint
for iOS.  Unfortunately this variable is often set at the environment
variable level on many Linux distributions so it can cause the Makefile
to think it is attempting to be cross-compiled when that is not the
intention.

This change modifies the Makefile so it attempts to cross compile only
when both HOST and HOSTROOT environment variable (which are both
required for cross compilation) are set.