macOS build tweaks
1. Building the DMG file now requires explicitly calling the 'make
   TuxPaint.dmg' target.

2. Messages output after completing a Makefile target has been
   customized on macOS.

The motivation here is Apple Silicon.  With the instruction of Apple
Silicon, the porting developer may need to sign the app bundle and/or
create the app bundle as universal before creating the DMG file, so the
Makefile no longer auto-creates the DMG file.  Instead, macOS-specific
messages have been added so the developer knows what to do next.