Brush tips - show when anim/rotate/dir, and spacing
Show tips when choosing brushes, to remind the user when they
are animated, and or rotational/directional.

Show tips when brush spacing is changed, specifying the approximate
spacing.  Output is fuzzy, to avoid being overly technical (e.g., it
will never say "2 345/1000 times as big as" or "1/103 the size of")