Remove trailing period in application summary
For the AppStream metadata (appdata.xml), there is a
rule saying that the summary shouldn’t end with a period
(as confirmed by ‘appstream-util validate-strict’).
For the .desktop file, *most* applications omit the period.

The rule is not language-dependent, so I have also
removed the period from all translations (.po files).