[Del] key to quickly access a small eraser
A keyboard shortcut is now available for quickly accessing
a small eraser.  Hold the [Del] key while clicking, and drag
around the picture to erase.  Release the key and/or mouse to
return to normal operation.

(Eventually, this could be utilized when we can support
the recognition of a drawing tablet stylus' "eraser" tip;
see https://sourceforge.net/p/tuxpaint/feature-requests/8/)