[X] instead of [Del] for eraser quick access
Apple keyboards have only a "Delete" key (which acts as Backspace),
so contortions with Fn key required to access the keysym I was
using for eraser quick access.  h/t Mark Kim for pointing it out.

Because of this use of a printable glyph key, and to solve the
issue of on screen keyboard and in-progress text input getting
clobbered while accessing eraser, preventing access to the
quick eraser when either Text or Label tool are selected.