Brush spacing improvements
Each brush's initial spacing (calculated, or loaded
from .dat file) is recorded.  When the spacing UI
is used to select a close, but not-precisely-the-same,
spacing option, the default is chosen ("nudged" to it).

h/t Pere for pointing this out (e.g., "aa_round_seethru_05.png")

Also, meter's lowest value has a sliver dark blue.