WIP - Crop whitespace from label text for better fit
For example, when using a font like OpenDyslexia which has a very
large line spacing, the text gets shrunk since the overall bitmap
is too big for the space on the buttons (compared to the contents).

Cropping the whitespace will allow the text to remain larger.

WIP - The method of cropping isn't great.  Text with descenders
appear higher than text without, causing terrible alignment.
So for now, the new `crop_surface()` code is not being used.

Also, replace some "#ifdef DEBUG printf(...) #endif" with
a simple "DEBUG_PRINTF(...)".  And other minor tweaks.