[WIP] Starter/Template scale/smear/bkgd color options
Now possible to set scaling and smearing options for
each starter or template image.  A ".dat" file can be
created which describes whether it's okay to scale and crop
the image's top/bottom and/or left/right, and if not both,
whether to smear the sides of the starter/template to the
edges of the canvas (the prior behavior, and the default),
or use a specific solid color background.

Once finished, will close
https://sourceforge.net/p/tuxpaint/feature-requests/190/

Still to do:
 + Document
 + Remove debugging printf()s (change to DEBUG_PRINTF()?)
 + Add files for starters & templates that ship with Tux Paint
 + Test!