Template/Starter options: Ability to specify gravity
e.g. if the aspect ratio of the template/starter is wider than
that of the canvas, and "autoscale=vertical" (or "=both")
is specified, before it would chop the top and bottom of the
image off, displaying the vertical center of it on the canvas.
Now you can ask it to gravitate towards the top or bottom, if they
have content that's of higher significance than the opposite side.

h/t Pere & Albert for the suggestion.

TODO - Document, set some suitable defaults for images we ship.