Modifying stamp rotation process: easier 0 degree
Rotation off of zero (0) degrees only occurs when dragging outside
the bounding box (rectangle around the width/height) of the stamp.

No longer warping mouse to the right of the stamp, to cause it to
default to zero degrees.  (The mouse being at the center of the
stamp now causes this, per the above change.)

Motivation - Placing stamps on mobile (touchscreen) was difficult
and non-intuitive.  Now, you can at least double-tap the same
spot on the screen to position, and then place (without rotation)
the stamp.  It's still not perfect, and I think a UI option to
enable/disable stamp rotation (leaving it off by default) is needed.