More tweaks to Stamp rotation step
Locking to 0-degrees when mouse is within a UI-button-sized
square centered around the stamp's center (h/t Pere for the idea).

Change mouse pointer to 'hand' when in the center (so at the start
of the rotation step), and to the 'circle arrow' rotation symbol
when actually rotating.

WIP (disabled for now) Drawing a XOR'd line out of the stamp's
center to the mouse, to show the angle you're rotating at.
(Needs more work to avoid leaving residuals on the screen.)