2-point perspective via angles
Rather than using straight horizontal & vertical, it now snaps
to lines with the same slope as the two vanishing points
(the horizon chosen by the user), and lines perpendicular to that.

Note - Still need guides during drag.  Also, there's a bug when
calculating slope that causes the line to go way off in the
opposite direction, which needs addressing.