appdata: Replace named display_length with constant values
The named values for <display_length> were removed in Appstream 1.0,
released recently.  In their place, use integer values.

For the <requires> block, attempt to encode the same reasonableness
restrictions as are enforced on the windowsize setting. The code
requires the height to be at least 480px and the width at least 500px,
but there is no way to specify width versus height, just shortest versus
longest edge, so I just require that the shortest edge is at least
500px.

For the <recommends> block, 768px is the suggested number in the
appstream specification for laptops and tablets.

https://github.com/ximion/appstream/commit/005c5f104d45953cd25f842b70dcd3d32466333b
https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-relations-display_length