Transparent erasers
Added round erasers that only expose 25% of the background each
time you click and apply the eraser.

Reworked how erasers are handled in the code to make it easier
to add this.

Reworked how eraser icons are drawn in the selector on the right.
(Squares and non-fuzzy, non-transparent circles are solid;
fuzzy circles are solid in the center with a dithered edge;
transparent circles are entirely dithered.)

TODO
 * Document the new eraser type
 * Remove a visual artfact (XOR outline) when using Ctrl-Z to Undo
 * Eraser icons turn from black to white when temporarily switching
   to certain other tools (e.g., after prompts like Quit or Print,
   or after using Save)