Mend issue of rotating brushes not drawing on single click
Rotating brushes would not draw anything if you simply
click and release with no motion in between (unless brush
spacing was set to the minimum).  It will now draw the
default "up" direction of the brush in that situation.
(h/t Pere for reporting)

Implemented by using a boolean that tracks whether there has
been any motion since a click; if not, the `brush_counter` is
forced to reset (to 999), to insist on `brush_draw()` blitting
something.  (It used to always ignore if a rotational brush,
to avoid accidentally drawing in the wrong direction on release.)