Word wrap button labels on soft hypens
If the UTF-8 sequence for the Unicode Soft Hyphen ("Non-printable U+00AD")
occurs in the text to be used as the label of a button, and the label
is still too wide after any wrapping done based on either space and/or a
visible hyphen ("-", U+002D), a visible hyphen will be introduced and
the label will be wrapped.

e.g. Norwegian Nynorsk "Regnbogesyklus"

Regnboge-
syklus

(rather than all crammed on one line and impossible to read)

h/t Karl for the idea.

Closes https://sourceforge.net/p/tuxpaint/bugs/288/