Update places that mentioned "--fullscreen" with no args
Since `parse.gperf` was added long ago, the ability to run Tux Paint
with "--fullscreen" command-line, or simply "fullscreen" in the
config. file, no longer exists.

It must be `--fullscreen=yes` (or `--fullscreen=native` as an
alternative to `--fullscreen=yes --native`).  It can also accept
`--fullscreen=no` as an alternative to `--windowed`.

Updating the following places

 * BASH tab completion
 * `tuxpaint --help` output
 * FAQ

h/t Karl for reporting

Closes https://sourceforge.net/p/tuxpaint/bugs/289/